Make V8::Access::Names get and set methods faster
We have an application with a page that uses a lot therubyracer
and profiling it shows a clear bottleneck on accessible_names.
In ruby 2.0 and 2.1, +respond_to?+ lists only public methods and
should be a lot faster than constructing accessible_names.

See https://gist.github.com/nono/9829448 for the profiling trace.